fixed tracks apicall crash: don't map long into int

+++ Spoti-bot/ApiModels/Track.cs
@@ -5,7 +5,7 @@ namespace Spoti_bot.ApiModels
     public class Track
     {
         public string Id { get; set; }
-        public int AddedByTelegramUserId { get; set; }
+        public long AddedByTelegramUserId { get; set; }
         public DateTimeOffset AddedAt { get; set; }
     }
 }
